"Disable Narration" default cfg change
Narration is a vanilla accessibility option that allows for those who need it to enable an audio screen reader while playing the game.
The UT option "Disable Narration" completely disables this option, preventing users from enabling it and toggling through its options (Shows "Not Available")
The current default option is "true", which means if a modpack dev includes UT in a modpack and does not switch this option to "false", then this option remains inaccessible to players, potentially preventing those who need this accessibility option from being able to fully enjoy the game.
By changing the default option to "false", modpack devs will have to consciously and deliberately disable it. Otherwise, it remains accessible to players.